test(cli): extract integration test helpers refs #72
